WIP: Update PostBuilder to allow quote posts in replies.
