+++ src/idunno.AtProto/PublicAPI.Unshipped.txt
@@ -1,1 +1,3 @@
-﻿#nullable enable
+#nullable enable
+idunno.AtProto.Labels.SelfLabels.AddLabel(idunno.AtProto.Labels.SelfLabel! label) -> void
+idunno.AtProto.Labels.SelfLabels.Contains(idunno.AtProto.Labels.SelfLabel! label) -> bool
