+++ src/idunno.Bluesky/PublicAPI.Unshipped.txt
@@ -1,2 +1,23 @@
 #nullable enable
-idunno.Bluesky.PostBuilder.Validate(System.ComponentModel.DataAnnotations.ValidationContext! validationContext) -> System.Collections.Generic.IEnumerable<System.ComponentModel.DataAnnotations.ValidationResult!>!
+idunno.Bluesky.PostBuilder.Add(char c) -> idunno.Bluesky.PostBuilder!
+idunno.Bluesky.PostBuilder.Add(char c, int repeatCount) -> idunno.Bluesky.PostBuilder!
+idunno.Bluesky.PostBuilder.Add(idunno.AtProto.Labels.SelfLabel! selfLabel) -> idunno.Bluesky.PostBuilder!
+idunno.Bluesky.PostBuilder.Add(idunno.AtProto.Repo.StrongReference! strongReference) -> idunno.Bluesky.PostBuilder!
+idunno.Bluesky.PostBuilder.Add(idunno.Bluesky.RichText.HashTag! hashTag) -> idunno.Bluesky.PostBuilder!
+idunno.Bluesky.PostBuilder.Add(idunno.Bluesky.RichText.Link! link) -> idunno.Bluesky.PostBuilder!
+idunno.Bluesky.PostBuilder.Add(idunno.Bluesky.RichText.Mention! mention) -> idunno.Bluesky.PostBuilder!
+idunno.Bluesky.PostBuilder.Add(string! s) -> idunno.Bluesky.PostBuilder!
+idunno.Bluesky.PostBuilder.Equals(idunno.Bluesky.PostBuilder? other) -> bool
+idunno.Bluesky.PostBuilder.IsValid() -> bool
+idunno.Bluesky.PostBuilder.Label(idunno.Bluesky.PostSelfLabels! labels) -> void
+idunno.Bluesky.PostBuilder.Quote(idunno.AtProto.Repo.StrongReference! strongReference) -> idunno.Bluesky.PostBuilder!
+idunno.Bluesky.PostBuilder.ReplyTo(idunno.AtProto.Repo.StrongReference! postReference, idunno.Bluesky.BlueskyAgent! agent) -> System.Threading.Tasks.Task<idunno.Bluesky.PostBuilder!>!
+idunno.Bluesky.PostBuilder.ReplyTo(idunno.Bluesky.ReplyReferences! replyReferences) -> idunno.Bluesky.PostBuilder!
+idunno.Bluesky.PostBuilder.ValidationErrors() -> System.Collections.Generic.IEnumerable<string!>!
+idunno.Bluesky.PostBuilder.WithText(string! text) -> idunno.Bluesky.PostBuilder!
+static idunno.Bluesky.PostBuilder.Add(idunno.Bluesky.PostBuilder! postBuilder, char c) -> idunno.Bluesky.PostBuilder!
+static idunno.Bluesky.PostBuilder.Add(idunno.Bluesky.PostBuilder! postBuilder, idunno.AtProto.Labels.SelfLabel! selfLabel) -> idunno.Bluesky.PostBuilder!
+static idunno.Bluesky.PostBuilder.Add(idunno.Bluesky.PostBuilder! postBuilder, idunno.AtProto.Repo.StrongReference! strongReference) -> idunno.Bluesky.PostBuilder!
+static idunno.Bluesky.PostBuilder.operator +(idunno.Bluesky.PostBuilder! postBuilder, idunno.AtProto.Labels.SelfLabel! selfLabel) -> idunno.Bluesky.PostBuilder!
+static idunno.Bluesky.PostBuilder.operator +(idunno.Bluesky.PostBuilder! postBuilder, idunno.AtProto.Repo.StrongReference! strongReference) -> idunno.Bluesky.PostBuilder!
+static idunno.Bluesky.PostBuilder.Quote(idunno.Bluesky.PostBuilder! postBuilder, idunno.AtProto.Repo.StrongReference! strongReference) -> idunno.Bluesky.PostBuilder!
